+++ src/test/java/domain/MovementTest.java
@@ -5,7 +5,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import domain.position.Direction;
-import domain.position.Movement;
+import domain.piece.path.Movement;
 import domain.position.Position;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
