+++ src/test/java/domain/piece/path/FixedMultiStepPathFinderTest.java
@@ -6,7 +6,6 @@ import static fixtures.PositionFixture.H0;
 import static org.assertj.core.api.Assertions.*;
 
 import domain.position.Direction;
-import domain.position.Movement;
 import domain.position.Position;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
